madden_discord_bot/utils/player_lookup.py
import aiohttp
import asyncio
from bs4 import BeautifulSoup
import re
import json

class PlayerLookup:
    def __init__(self):
        self.session = None
        # Cache for player data to avoid repeated lookups
        self.player_cache = {}
    
    async def get_session(self):
        if self.session is None:
            self.session = aiohttp.ClientSession()
        return self.session
    
    async def close_session(self):
        if self.session:
            await self.session.close()
            self.session = None
    
    async def lookup_player_data(self, player_name):
        """Lookup player data from EA Sports or other sources"""
        # Check cache first
        cache_key = player_name.lower().strip()
        if cache_key in self.player_cache:
            return self.player_cache[cache_key]
        
        try:
            session = await self.get_session()
            
            # Try EA Sports Madden website first
            player_data = await self._search_ea_madden(session, player_name)
            
            if not player_data:
                # Fallback to estimated data based on common knowledge
                player_data = self._estimate_player_data(player_name)
            
            # Cache the result
            self.player_cache[cache_key] = player_data
            return player_data
            
        except Exception as e:
            print(f"Error looking up {player_name}: {e}")
            return self._estimate_player_data(player_name)
    
    async def _search_ea_madden(self, session, player_name):
        """Search EA Madden website for player data"""
        try:
            # This is a simplified version - EA's site might require more complex scraping
            search_url = f"https://www.ea.com/games/madden-nfl/madden-nfl-25/ratings"
            
            headers = {
                'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
            }
            
            # For now, we'll return None and rely on OpenAI + estimation
            # In production, you'd implement the actual EA scraping here
            return None
            
        except Exception as e:
            print(f"EA lookup error for {player_name}: {e}")
            return None
    
    def _estimate_player_data(self, player_name):
        """Estimate player data based on known players"""
        # This is a fallback with some common players
        # In production, you'd have a more comprehensive database
        
        name_lower = player_name.lower()
        
        # Star QBs
        if any(name in name_lower for name in ['mahomes', 'josh allen', 'burrow', 'lamar', 'dak']):
            return {'ovr': 95, 'age': 27, 'position': 'qb', 'dev': 'superstar'}
        
        # Elite WRs
        elif any(name in name_lower for name in ['jefferson', 'hill', 'adams', 'kupp', 'diggs']):
            return {'ovr': 96, 'age': 26, 'position': 'wr', 'dev': 'superstar'}
        
        # Default estimates
        else:
            return {'ovr': 80, 'age': 26, 'position': 'wr', 'dev': 'normal'}

# Global instance
player_lookup = PlayerLookup()

