madden_discord_bot/requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
openai>=1.0.0
aiohttp>=3.8.0
beautifulsoup4>=4.9.0
